+++ src/main/java/dev/lambdaurora/mcdev/api/ClassTweakerToTransformer.java
@@ -8,8 +8,10 @@
 
 package dev.lambdaurora.mcdev.api;
 
-import net.fabricmc.accesswidener.AccessWidenerReader;
-import net.fabricmc.accesswidener.AccessWidenerVisitor;
+import net.fabricmc.classtweaker.api.ClassTweakerReader;
+import net.fabricmc.classtweaker.api.visitor.AccessWidenerVisitor;
+import net.fabricmc.classtweaker.api.visitor.ClassTweakerVisitor;
+import org.jspecify.annotations.NullMarked;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -21,49 +23,63 @@ import java.util.List;
 import java.util.Map;
 
 // This is an unsafe converter.
-public final class AccessWidenerToTransformer implements AccessWidenerVisitor {
+@NullMarked
+public final class ClassTweakerToTransformer implements ClassTweakerVisitor {
 	private final Map<String, Boolean> classes = new HashMap<>();
 	private final Map<FieldEntry, Boolean> fields = new HashMap<>();
 	private final Map<MethodEntry, Boolean> methods = new HashMap<>();
 
 	@Override
-	public void visitClass(String name, AccessWidenerReader.AccessType access, boolean transitive) {
-		this.classes.compute(
-				name, (key, value) -> value == null
-						? this.getAccess(access)
-						: (value || this.getAccess(access))
-		);
+	public AccessWidenerVisitor visitAccessWidener(String owner) {
+		return new AccessWidenerToTransformer(owner);
 	}
 
-	@Override
-	public void visitField(
-			String owner, String name, String descriptor, AccessWidenerReader.AccessType access, boolean transitive
-	) {
-		this.fields.compute(
-				new FieldEntry(owner, name),
-				(key, value) -> value == null
-						? this.getAccess(access)
-						: (value || this.getAccess(access))
-		);
-	}
+	private class AccessWidenerToTransformer implements AccessWidenerVisitor {
+		private final String owner;
 
-	@Override
-	public void visitMethod(
-			String owner, String name, String descriptor, AccessWidenerReader.AccessType access, boolean transitive
-	) {
-		this.methods.compute(
-				new MethodEntry(owner, name, descriptor),
-				(key, value) -> value == null
-						? this.getAccess(access)
-						: (value || this.getAccess(access))
-		);
-	}
+		public AccessWidenerToTransformer(String owner) {
+			this.owner = owner;
+		}
 
-	private boolean getAccess(AccessWidenerReader.AccessType accessType) {
-		return switch (accessType) {
-			case ACCESSIBLE -> false;
-			case EXTENDABLE, MUTABLE -> true;
-		};
+		@Override
+		public void visitClass(AccessType access, boolean transitive) {
+			classes.compute(
+					this.owner, (key, value) -> value == null
+							? this.getAccess(access)
+							: (value || this.getAccess(access))
+			);
+		}
+
+		@Override
+		public void visitField(
+				String name, String descriptor, AccessType access, boolean transitive
+		) {
+			fields.compute(
+					new FieldEntry(this.owner, name),
+					(key, value) -> value == null
+							? this.getAccess(access)
+							: (value || this.getAccess(access))
+			);
+		}
+
+		@Override
+		public void visitMethod(
+				String name, String descriptor, AccessType access, boolean transitive
+		) {
+			methods.compute(
+					new MethodEntry(this.owner, name, descriptor),
+					(key, value) -> value == null
+							? this.getAccess(access)
+							: (value || this.getAccess(access))
+			);
+		}
+
+		private boolean getAccess(AccessType accessType) {
+			return switch (accessType) {
+				case ACCESSIBLE -> false;
+				case EXTENDABLE, MUTABLE -> true;
+			};
+		}
 	}
 
 	public List<String> toLines() {
@@ -94,10 +110,10 @@ public final class AccessWidenerToTransformer implements AccessWidenerVisitor {
 	}
 
 	public static void convert(Path source, Path dest) throws IOException {
-		var converter = new AccessWidenerToTransformer();
+		var converter = new ClassTweakerToTransformer();
 
 		try (var reader = Files.newBufferedReader(source, StandardCharsets.UTF_8)) {
-			var awReader = new AccessWidenerReader(converter);
+			var awReader = ClassTweakerReader.create(converter);
 			awReader.read(reader);
 		}